backend/requirements.txt
fastapi==0.136.0
uvicorn==0.45.0
pydantic-ai==1.85.0
python-dotenv==1.2.2
pandas==3.0.2